refactor: small updates
Add JSDoc types and comments.

Split code.

Rename `build` to `buildServer`.

Rename `pathToIndex` to `projectRootPath`.

Change `npm run dev` to `npm start` in the `directories.js` comment.
