+++ server/src/config/directories.js
@@ -1,13 +1,13 @@
-import { join } from "path";
+import path from "path";
 import process from "process";
 
 // Checks if the directory is explicitly specified.
-// It can be specified by writing for example `npm run dev other-frameworks-directory`.
+// It can be specified by writing for example `npm start other-frameworks-directory`.
 const isFrameworksDirectorySpecifies = process.argv.length === 3;
 
 const frameworksDirectory = isFrameworksDirectorySpecifies
-  ? join(process.cwd(), "..", process.argv[2])
-  : join(process.cwd(), "..", "frameworks");
+  ? path.join(process.cwd(), "..", process.argv[2])
+  : path.join(process.cwd(), "..", "frameworks");
 
 if (isFrameworksDirectorySpecifies) {
   console.log(`Changing working directory to ${process.argv[2]}`);
