+++ server/src/csp/cspRouter.js
@@ -3,13 +3,12 @@ import { addCSP, disableCSP, enableCSP, getCSP } from "./cspControllers.js";
 /**
  * A plugin that provide encapsulated routes
  * @param {import("fastify").FastifyInstance} fastify
- * @param {import("fastify").RegisterOptions} options
  */
-async function routes(fastify, _options) {
+async function routes(fastify) {
   fastify.addContentTypeParser(
     "application/csp-report",
     { parseAs: "string" },
-    fastify.getDefaultJsonParser("ignore", "ignore")
+    fastify.getDefaultJsonParser("ignore", "ignore"),
   );
 
   fastify.get("/enable", enableCSP);
