+++ server/src/frameworks/frameworksRouter.js
@@ -1,27 +1,11 @@
-import { loadFrameworkVersionInformation } from "./frameworksControllers.js";
+import { getFrameworksVersions } from "./frameworksControllers.js";
 
 /**
  * A plugin that provide encapsulated routes
  * @param {import("fastify").FastifyInstance} fastify
- * @param {import("fastify").RegisterOptions} options
  */
-async function routes(fastify, _options) {
-  fastify.get("/ls", async function (_request, reply) {
-    performance.mark("Start");
-
-    const frameworks = await loadFrameworkVersionInformation();
-    reply.send(frameworks);
-
-    performance.mark("End");
-    const executionTime = performance.measure(
-      "/ls duration measurement",
-      "Start",
-      "End",
-    ).duration;
-
-    console.log(`/ls duration: ${executionTime}ms`);
-    return;
-  });
+async function routes(fastify) {
+  fastify.get("/ls", getFrameworksVersions);
 }
 
 export default routes;
